+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This is a changelog for Piwik platform developers. All changes for our HTTP API'
 
 ### Breaking Changes
 * The method `Dimension::getId()` has been set as `final`. It is not allowed to overwrite this method.
+* The segment operators `=@` (contains) and `!@` (does not contain) are now case insenstive.
 
 ### Deprecations